Changed display to window, refactored and added nodes
Big problem here. Windows are created BUT they're not shown until after the function ends, which is never if the script kills the windows to avoid cluttering the screen. This needs a fix, but I also need to sleep and only I am using the plugin right now, so there it stays for now.